adds unittests for DescriptionInput
